feat: add distinct property to handlers
