+++ lib/flutter_micro_app.dart
@@ -1,12 +1,13 @@
 library flutter_micro_app;
 
-export 'src/controllers/micro_app_event_controller.dart';
+export 'src/controllers/app_event/micro_app_event_controller.dart';
 export 'src/controllers/navigator_controller.dart';
 export 'src/core/micro_app.dart';
 export 'src/core/micro_host.dart';
 export 'src/core/micro_page.dart';
 export 'src/core/micro_widget.dart';
 export 'src/entities/events/micro_app_event.dart';
+export 'src/entities/events/micro_app_event_handler.dart';
 export 'src/entities/micro_app_preferences.dart';
 export 'src/entities/router/base_route.dart';
 export 'src/presentation/widgets/micro_widget_builder.dart';
