+++ lib/src/controllers/app_event/micro_app_event_helper.dart
@@ -1,128 +1,15 @@
-// ignore_for_file: cancel_subscriptions
 import 'dart:async';
 
-import 'package:flutter/services.dart';
-import 'package:flutter_micro_app/flutter_micro_app.dart';
-import 'package:flutter_micro_app/src/services/native_service.dart';
+import '../../entities/events/micro_app_event.dart';
+import '../../entities/events/micro_app_event_handler.dart';
 
-/// [MicroAppEventHandler]
-class MicroAppEventHandler<T> extends EventChannelsEquatable {
-  final String? id;
-  final MicroAppEventOnEvent onEvent;
-  final MicroAppEventOnDone? onDone;
-  final MicroAppEventOnError? onError;
-  final bool? cancelOnError;
-
-  const MicroAppEventHandler(
-    this.onEvent, {
-    this.onDone,
-    this.onError,
-    this.cancelOnError,
-    List<String> channels = const [],
-    this.id,
-  }) : super(channels);
-
-  @override
-  List<Object?> get props =>
-      [onEvent, onDone, onError, cancelOnError, channels, id];
-}
-
-/// [MicroAppEventController]
-class MicroAppEventController {
-  final String _channel = Constants.channelAppEvents;
-  final StreamController<MicroAppEvent> _controller =
-      StreamController.broadcast();
-  late MicroAppNativeService _microAppNativeService;
-  late HandlerRegisterHelper _handlerRegisterHelper;
-
-  /// Used to listen all micro app events
-  MicroAppEventSubscription get onEvent => _controller.stream.listen;
-
-  /// Can be used to filter events and listen to app events, anywhere
-  Stream<MicroAppEvent> get stream => _controller.stream;
-  bool get hasSubscribers => _controller.hasListener;
-  bool get hasHandlers => _handlers.isNotEmpty;
-
-  final Map<MicroAppEventHandler, StreamSubscription<MicroAppEvent>> _handlers =
-      {};
-
-  /// get all handlers and subscriptions availables
-  Map<MicroAppEventHandler, StreamSubscription<MicroAppEvent>> get handlers =>
-      _handlers;
-
-  /// Factory instance
-  factory MicroAppEventController() => instance;
-
-  /// MicroAppController instance
-  static MicroAppEventController instance = MicroAppEventController._();
-  MicroAppEventController._() {
-    _handlerRegisterHelper = HandlerRegisterHelper();
-    _microAppNativeService = MicroAppNativeService(_channel,
-        methodCallHandler: (MethodCall call) async {
-      _controller
-          .add(MicroAppEvent(name: call.method, payload: call.arguments));
-    });
-  }
-
-  /// ⚠️ It is used only for unit tests purposes
-  factory MicroAppEventController.$testOnlyPurpose() =>
-      MicroAppEventController._();
-
-  /// [MicroAppEvent]
-  void emit(MicroAppEvent event) {
-    if (MicroAppPreferences.config.nativeEventsEnabled) {
-      _microAppNativeService.emit(
-          Constants.methodMicroAppEvent, event.toString());
-    }
-    _controller.add(event);
-  }
-
-  /// register handler
-  StreamSubscription<MicroAppEvent>? registerHandler<T>(
-      MicroAppEventHandler? handler) {
-    if (handler == null) return null;
-    final subscription =
-        _handlerRegisterHelper.registerHandler(stream, handler);
-    _handlers.putIfAbsent(handler, () => subscription);
-    return subscription;
-  }
-
-  /// pauseAllHandlers
-  void pauseAllHandlers() =>
-      _handlerRegisterHelper.pauseAllSubscriptions(_handlers);
-
-  /// resumeAllHandlers
-  void resumeAllHandlers() =>
-      _handlerRegisterHelper.resumeAllSubscriptions(_handlers);
-
-  /// unregisterAllHandlers
-  Future<void> unregisterAllHandlers() =>
-      _handlerRegisterHelper.unregisterAllSubscriptions(_handlers);
-
-  /// unregisterHandler
-  Future<void> unregisterHandler({String? id, List<String>? channels}) =>
-      _handlerRegisterHelper.unregisterSubscription(_handlers,
-          id: id, channels: channels);
-
-  /// hasHandler
-  bool hasHandler({String? id, List<String>? channels}) =>
-      _handlerRegisterHelper
-          .findHandlerEntries(handlers, id: id, channels: channels)
-          .isNotEmpty;
-
-  /// dispose controller (do not dispose this, if there is not a very specific situation case)
-  void dispose() {
-    _controller.close();
-  }
-}
-
-class HandlerRegisterHelper {
+class MicroAppEventHelper {
   /// registerHandler
   StreamSubscription<MicroAppEvent> registerHandler(
       Stream<MicroAppEvent> stream, MicroAppEventHandler handler) {
     return stream
-        // TODO: implementar distinct
-        // .distinct((previous, event) => event.distinct)
+        .distinct((previousEvent, currentEvent) =>
+            handleDistinct(handler, previousEvent, currentEvent))
         .where((event) => handlerHasSameEventTypeOrDynamic(handler, event))
         .where((event) => containsSomeChannelsOrHandlerHasNoChannels(
             handler.channels, event.channels))
@@ -201,6 +88,12 @@ class HandlerRegisterHelper {
     }
 
     return entries;
+  }
+
+  bool handleDistinct(MicroAppEventHandler handler, MicroAppEvent previousEvent,
+      MicroAppEvent currentEvent) {
+    if (handler.distinct == false) return false;
+    return !(handler.distinct && previousEvent != currentEvent);
   }
 
   bool handlerHasSameEventTypeOrDynamic(
